+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release
+* Features:
+  * USB printer support for Android
+  * Methods for discovering USB devices
+  * Support for ESC/POS commands
+  * Raw data printing
+  * Text printing with formatting